Fix/Improve: Modify drag code to allow moving elements
IsInside using the original mouse-down coordinates is not a good test to
determine whether dragging behaviour is related to a particular element.
This is because it is often desirable to move a clicked element with the
drag motion, which could move it clear of the original mouse-down
location and cause it to fail the IsInside condition and break out of
the drag. Instead, we should store which element was originally clicked.